# storage/api-docs/api-docs.yaml
...
openapi: 3.0.0
info:
  title: Development Challenge for PHP
  description: This is a sample API documentation for the Development Challenge for
    PHP
  version: 1.0.0
paths:
...
  "/api/directors":
    get:
      tags:
      - Directors
      summary: Display a listing of the directors
      operationId: '0232862a597e188b8d6b90bcff38654f'
      responses:
        '200':
          description: A listing of the directors
          content:
            application/json:
              schema:
                properties:
                  directors:
                    "$ref": "#/components/schemas/Director"
                type: object
    post:
      tags:
      - Directors
      summary: Store a newly created resource in storage
      operationId: 745ff928efedfec1f1958d43a96aeade
      requestBody:
        description: Pass director data
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                genre:
                  type: string
                birth_date:
                  type: string
              type: object
      responses:
        '200':
          description: Director created successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  director:
                    "$ref": "#/components/schemas/Director"
                type: object
        '422':
          description: Invalid data passed
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                type: object
  "/api/directors/{director}":
    get:
      tags:
      - Directors
      summary: Display the specified director
      operationId: ab3e1e4496f1c4d99a772ddcfd06547d
      parameters:
      - name: director
        in: path
        description: Director ID
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Director details
          content:
            application/json:
              schema:
                properties:
                  director:
                    "$ref": "#/components/schemas/Director"
                type: object
        '404':
          description: Director not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
    put:
      tags:
      - Directors
      summary: Update the specified director in storage
      operationId: 2e8b01ff962d16043e3340cbb2f6d874
      parameters:
      - name: director
        in: path
        description: Director ID
        required: true
        schema:
          type: integer
      requestBody:
        description: Pass director data
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                genre:
                  type: string
                birth_date:
                  type: string
              type: object
      responses:
        '200':
          description: Director updated successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  director:
                    "$ref": "#/components/schemas/Director"
                type: object
        '404':
          description: Director not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
    delete:
      tags:
      - Directors
      summary: Remove the specified director from storage
      operationId: 9af1d1420510feb4e4ff4dad684204b4
      parameters:
      - name: director
        in: path
        description: Director ID
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Director deleted successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  director:
                    "$ref": "#/components/schemas/Director"
                type: object
        '403':
          description: Director has movies or tv shows or episodes
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
        '404':
          description: Director not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
...
components:
  schemas:
    GenderEnum:
      description: 'Gender options: Male, Female'
      type: string
      enum:
      - M
      - F
...
    Director:
      title: Director
      description: Director model
      properties:
        id:
          description: Director id
          type: integer
          example: '1'
        name:
          description: Director name
          type: string
          example: John Doe
        birth_date:
          description: Director birth date
          type: string
          format: date
          example: '1990-01-01'
        gender:
          "$ref": "#/components/schemas/GenderEnum"
      type: object
...